Enable configuration of hc5 connectionKeepAlive and maxConnectionsPerRoute
Why:
the rootless podman api server has a low idle timeout, by default 10 seconds, so we have been encountering
exceptions when we have multiple pooled connections that idle longer than this in the client.
Then the next request leases one connection, which fails because the server has closed the connection
and then the httpclient retries once, and fails again on the next connection it leases. This exception
propagates up to the caller.

By configuring a lower connectionKeepAlive we should be able to prevent old connections being leased
from the connection pool. Configuring maxConnectionsPerRoute could be another safeguard since it would
prevent there being multiple half-open connections in the pool for a route.

For backwards compatibility if maxConnectionsPerRoute is not set then we set it to be equal to maxConnections